(1) rename utils to common_utils (2) move metrics to nn/
